libgimpconfig: implement serializing GimpParasite into a GimpConfig.
Used for instance if a plug-in had a GimpParasite argument (or aux
argument), which might not seem too useful at first. But anyway we
propose the type, so let's support it properly.

And actually I am going to use it in the next commit as a trick to save
binary data in stored last values for a plug-in.