libgimpbase: fix gimp_parasite_new() annotation.
Without the (element-type), parasite cannot be created in at least
PyGObject binding (I have not tested parasite creation in other
bindings) and we end up with such error:

> ValueError: Pointer arguments are restricted to integers, capsules, and None. See: https://bugzilla.gnome.org/show_bug.cgi?id=683599

With this change, we can create it from a list of ints (e.g. for a text,
encode it into UTF-8 bytes).