plug-ins: fix #6723 Unable to open DDS files.
This is the same as what we did for #5357
but in the reverse situation. The reason
for not doing the same at that time was
because I wasn't sure if that was valid
in all cases.

Looking at the documentation it does seem
to be the right solution and is working for
the supplied example image.

(cherry picked from commit 48ff5313c3c7edc360d19db202b9c0398c1793f4)

# Conflicts:
#	plug-ins/file-dds/ddsread.c