gitlab-ci: store flatpak-build output into a log file.
Without this, the flatpak build is just too long and Gitlab CI just
stops logging it. So we end up with a log saying the quite useless (at
the end):

> Job's log exceeded limit of 16777216 bytes.
> Job execution will continue but no more output will be collected.

We don't even reach babl/GEGL/GIMP builds, which are the most important.
With this little trick, I am redirecting output to a log file and simply
including said log into the artifacts, hence allowing to debug the full
build.