gitlab-ci: add support for flatpak and Windows installer jobs in MR…
… based on set labels.

If the label "5. Windows Installer" is set, the MR pipeline should
trigger the Windows installer pipeline as well.

If the label "5. Flatpak package" is set, it should generate the flatpak
(not publish it obviously, yet it can be downloaded and installed
manually).

This will allow us to easily test MRs and allow people to test our code
before we merge it to the main branches.