devel-docs: some info on new Gitlab MR feature to create test packages.
How to create the packages through MR labels and where to find the
resulting test packages…

Note that it would be nice if we had automatic messages on the MR
writing down the procedure (with generated links for the specific
pipeline ID) once a pipeline succeeds.
It would simplify the whole process even further. We can see some other
day how such a thing could be automated.