NEWS: adding smooth scroll coming with GTK3.
Something which came with the 2.99.2 release already, though I update
the NEWS file retroactively because we forgot to write this down and
that's a pretty important feature (for ctrl-scroll zooming for instance,
which became a lot more useful in GIMP 2.99.x). This should not be
forgotten when we'll write the GIMP 3 release note, so I'm adding it in
this file where we'll get most contents from.