meson.build: fix more xmllint validation.
The authors.xml validation was also not run. This is nearly the last of
getting rid of run_target(). There is still the desktop file validation
but it doesn't have any output argument. We'll see how we update this
last one.

The only other usages of run_target() are proper usage (creating
'install-*' targets).