app: fix wrong unref-ing.
Testing this middle-click opening of image by their PATH/URI in the
toolbox, I realized there was a bug. The original author was obviously
intending to unref the toolbox which was ref-ed when calling
gtk_clipboard_request_text(), not freeing the toolbox context, which
could have dire consequences!

Fixes this CRITICAL:

> g_object_unref: assertion 'G_IS_OBJECT (object)' failed